refactor: factor out discovery method for easier API
