+++ index.js
@@ -1,8 +1,10 @@
-const { LoupedeckDevice, LoupedeckLive, LoupedeckLiveS } = require('./device')
+const { LoupedeckLive, LoupedeckLiveS } = require('./device')
 const { HAPTIC } = require('./constants')
+const { discover, listDevices } = require('./discovery')
 
 module.exports = {
-    LoupedeckDevice,
+    discover,
+    listDevices,
     LoupedeckLive,
     LoupedeckLiveS,
     HAPTIC
